dt to json cli
