+++ tests/a_scripts/data_prep/test_convert_delta_to_json.py
@@ -9,7 +9,7 @@ from argparse import Namespace
 from typing import Any
 from unittest.mock import MagicMock, mock_open, patch
 
-from scripts.data_prep.convert_delta_to_json import (
+from llmfoundry.data_prep.convert_delta_to_json import (
     download,
     fetch_DT,
     format_tablename,
@@ -20,11 +20,11 @@ from scripts.data_prep.convert_delta_to_json import (
 
 class TestConvertDeltaToJsonl(unittest.TestCase):
 
-    @patch('scripts.data_prep.convert_delta_to_json.sql.connect')
-    @patch('scripts.data_prep.convert_delta_to_json.os.makedirs')
-    @patch('scripts.data_prep.convert_delta_to_json.iterative_combine_jsons')
-    @patch('scripts.data_prep.convert_delta_to_json.fetch')
-    @patch('scripts.data_prep.convert_delta_to_json.WorkspaceClient')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.sql.connect')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.os.makedirs')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.iterative_combine_jsons')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.fetch')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.WorkspaceClient')
     def test_stream_delta_to_json(
         self,
         mock_workspace_client: Any,
@@ -66,7 +66,7 @@ class TestConvertDeltaToJsonl(unittest.TestCase):
             '/path/to/jsonl/combined.jsonl',
         )
 
-    @patch('scripts.data_prep.convert_delta_to_json.os.listdir')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.os.listdir')
     @patch(
         'builtins.open',
         new_callable=mock_open,
@@ -102,7 +102,7 @@ class TestConvertDeltaToJsonl(unittest.TestCase):
         """
         self.assertEqual(mock_file().write.call_count, 2)
 
-    @patch('scripts.data_prep.convert_delta_to_json.SparkSession')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.SparkSession')
     def test_run_query_dbconnect(self, mock_spark: Any):
         method = 'dbconnect'
         mock_cursor = None
@@ -118,7 +118,7 @@ class TestConvertDeltaToJsonl(unittest.TestCase):
         mock_spark.sql.assert_called_once_with('SELECT * FROM table')
         self.assertEqual(result, 'result')
 
-    @patch('scripts.data_prep.convert_delta_to_json.Cursor')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.Cursor')
     def test_run_query_dbsql(self, mock_cursor: Any):
         method = 'dbsql'
         mock_cursor.fetchall.return_value = 'result'
@@ -134,14 +134,14 @@ class TestConvertDeltaToJsonl(unittest.TestCase):
         mock_cursor.execute.assert_called_once_with('SELECT * FROM table')
         self.assertEqual(result, 'result')
 
-    @patch('scripts.data_prep.convert_delta_to_json.requests.get')
-    @patch('scripts.data_prep.convert_delta_to_json.pd.DataFrame.to_json')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.requests.get')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.pd.DataFrame.to_json')
     @patch(
-        'scripts.data_prep.convert_delta_to_json.os.path.join',
+        'llmfoundry.data_prep.convert_delta_to_json.os.path.join',
         return_value='/fake/path/part_1.jsonl',
     )
     @patch(
-        'scripts.data_prep.convert_delta_to_json.time.sleep',
+        'llmfoundry.data_prep.convert_delta_to_json.time.sleep',
     )  # Mock sleep to speed up the test
     def test_download_success(
         self,
@@ -174,12 +174,12 @@ class TestConvertDeltaToJsonl(unittest.TestCase):
 
         mock_get.assert_called_once_with('http://fakeurl.com/data')
 
-    @patch('scripts.data_prep.convert_delta_to_json.sql.connect')
-    @patch('scripts.data_prep.convert_delta_to_json.DatabricksSession')
-    @patch('scripts.data_prep.convert_delta_to_json.WorkspaceClient')
-    @patch('scripts.data_prep.convert_delta_to_json.os.makedirs')
-    @patch('scripts.data_prep.convert_delta_to_json.iterative_combine_jsons')
-    @patch('scripts.data_prep.convert_delta_to_json.fetch')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.sql.connect')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.DatabricksSession')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.WorkspaceClient')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.os.makedirs')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.iterative_combine_jsons')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.fetch')
     def test_dbconnect_called(
         self,
         mock_fetch: Any,
@@ -216,12 +216,12 @@ class TestConvertDeltaToJsonl(unittest.TestCase):
             cluster_id=args.cluster_id,
         )
 
-    @patch('scripts.data_prep.convert_delta_to_json.sql.connect')
-    @patch('scripts.data_prep.convert_delta_to_json.DatabricksSession')
-    @patch('scripts.data_prep.convert_delta_to_json.WorkspaceClient')
-    @patch('scripts.data_prep.convert_delta_to_json.os.makedirs')
-    @patch('scripts.data_prep.convert_delta_to_json.iterative_combine_jsons')
-    @patch('scripts.data_prep.convert_delta_to_json.fetch')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.sql.connect')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.DatabricksSession')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.WorkspaceClient')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.os.makedirs')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.iterative_combine_jsons')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.fetch')
     def test_sqlconnect_called_dbr13(
         self,
         mock_fetch: Any,
@@ -253,12 +253,12 @@ class TestConvertDeltaToJsonl(unittest.TestCase):
             access_token=args.DATABRICKS_TOKEN,
         )
 
-    @patch('scripts.data_prep.convert_delta_to_json.sql.connect')
-    @patch('scripts.data_prep.convert_delta_to_json.DatabricksSession')
-    @patch('scripts.data_prep.convert_delta_to_json.WorkspaceClient')
-    @patch('scripts.data_prep.convert_delta_to_json.os.makedirs')
-    @patch('scripts.data_prep.convert_delta_to_json.iterative_combine_jsons')
-    @patch('scripts.data_prep.convert_delta_to_json.fetch')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.sql.connect')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.DatabricksSession')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.WorkspaceClient')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.os.makedirs')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.iterative_combine_jsons')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.fetch')
     def test_sqlconnect_called_dbr14(
         self,
         mock_fetch: Any,
@@ -290,12 +290,12 @@ class TestConvertDeltaToJsonl(unittest.TestCase):
             access_token=args.DATABRICKS_TOKEN,
         )
 
-    @patch('scripts.data_prep.convert_delta_to_json.sql.connect')
-    @patch('scripts.data_prep.convert_delta_to_json.DatabricksSession')
-    @patch('scripts.data_prep.convert_delta_to_json.WorkspaceClient')
-    @patch('scripts.data_prep.convert_delta_to_json.os.makedirs')
-    @patch('scripts.data_prep.convert_delta_to_json.iterative_combine_jsons')
-    @patch('scripts.data_prep.convert_delta_to_json.fetch')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.sql.connect')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.DatabricksSession')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.WorkspaceClient')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.os.makedirs')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.iterative_combine_jsons')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.fetch')
     def test_sqlconnect_called_https(
         self,
         mock_fetch: Any,
@@ -327,12 +327,12 @@ class TestConvertDeltaToJsonl(unittest.TestCase):
             access_token=args.DATABRICKS_TOKEN,
         )
 
-    @patch('scripts.data_prep.convert_delta_to_json.sql.connect')
-    @patch('scripts.data_prep.convert_delta_to_json.DatabricksSession')
-    @patch('scripts.data_prep.convert_delta_to_json.WorkspaceClient')
-    @patch('scripts.data_prep.convert_delta_to_json.os.makedirs')
-    @patch('scripts.data_prep.convert_delta_to_json.iterative_combine_jsons')
-    @patch('scripts.data_prep.convert_delta_to_json.fetch')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.sql.connect')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.DatabricksSession')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.WorkspaceClient')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.os.makedirs')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.iterative_combine_jsons')
+    @patch('llmfoundry.data_prep.convert_delta_to_json.fetch')
     def test_serverless(
         self,
         mock_fetch: Any,
